Crate google_clouddebugger2

This documentation was generated from Cloud Debugger crate version 1.0.9+20190614, where 20190614 is the exact revision of the clouddebugger:v2 schema built by the mako code generator v1.0.9.

Everything else about the Cloud Debugger v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

controller

debuggees breakpoints list, debuggees breakpoints update and debuggees register

debugger

debuggees breakpoints delete, debuggees breakpoints get, debuggees breakpoints list, debuggees breakpoints set and debuggees list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.debugger().debuggees_breakpoints_set(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-clouddebugger2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_clouddebugger2 as clouddebugger2;
use clouddebugger2::Breakpoint;
use clouddebugger2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use clouddebugger2::CloudDebugger;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudDebugger::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Breakpoint::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.debugger().debuggees_breakpoints_set(req, "debuggeeId")
             .client_version("sed")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AliasContext
An alias to a repo revision.

Breakpoint
Represents the breakpoint specification, status and results.

Chunk

CloudDebugger
Central instance to access all CloudDebugger related resource activities

CloudRepoSourceContext
A CloudRepoSourceContext denotes a particular revision in a cloud repo (a repo hosted by the Google Cloud Platform).

CloudWorkspaceId
A CloudWorkspaceId is a unique identifier for a cloud workspace. A cloud workspace is a place associated with a repo where modified files can be stored before they are committed.

CloudWorkspaceSourceContext
A CloudWorkspaceSourceContext denotes a workspace at a particular snapshot.

ContentRange
Implements the Content-Range header, for serialization only

ControllerDebuggeeBreakpointListCall
Returns the list of all active breakpoints for the debuggee.

ControllerDebuggeeBreakpointUpdateCall
Updates the breakpoint state or mutable fields. The entire Breakpoint message must be sent back to the controller service.

ControllerDebuggeeRegisterCall
Registers the debuggee with the controller service.

ControllerMethods
A builder providing access to all methods supported on controller resources. It is not used directly, but through the CloudDebugger hub.

Debuggee
Represents the debugged application. The application may include one or more replicated processes executing the same code. Each of these processes is attached with a debugger agent, carrying out the debugging commands. Agents attached to the same debuggee identify themselves as such by using exactly the same Debuggee message value when registering.

DebuggerDebuggeeBreakpointDeleteCall
Deletes the breakpoint from the debuggee.

DebuggerDebuggeeBreakpointGetCall
Gets breakpoint information.

DebuggerDebuggeeBreakpointListCall
Lists all breakpoints for the debuggee.

DebuggerDebuggeeBreakpointSetCall
Sets the breakpoint to the debuggee.

DebuggerDebuggeeListCall
Lists all the debuggees that the user has access to.

DebuggerMethods
A builder providing access to all methods supported on debugger resources. It is not used directly, but through the CloudDebugger hub.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ExtendedSourceContext
An ExtendedSourceContext is a SourceContext combined with additional details describing the context.

FormatMessage
Represents a message with parameters.

GerritSourceContext
A SourceContext referring to a Gerrit project.

GetBreakpointResponse
Response for getting breakpoint information.

GitSourceContext
A GitSourceContext denotes a particular revision in a third party Git repository (e.g. GitHub).

JsonServerError
A utility type which can decode a server response that indicates error

ListActiveBreakpointsResponse
Response for listing active breakpoints.

ListBreakpointsResponse
Response for listing breakpoints.

ListDebuggeesResponse
Response for listing debuggees.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ProjectRepoId
Selects a repo using a Google Cloud Platform project ID (e.g. winged-cargo-31) and a repo name within that project.

RangeResponseHeader

RegisterDebuggeeRequest
Request to register a debuggee.

RegisterDebuggeeResponse
Response for registering a debuggee.

RepoId
A unique identifier for a cloud repo.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

SetBreakpointResponse
Response for setting a breakpoint.

SourceContext
A SourceContext is a reference to a tree of files. A SourceContext together with a path point to a unique revision of a single file or directory.

SourceLocation
Represents a location in the source code.

StackFrame
Represents a stack frame context.

StatusMessage
Represents a contextual status message. The message can indicate an error or informational status, and refer to specific parts of the containing object. For example, the Breakpoint.status field can indicate an error referring to the BREAKPOINT_SOURCE_LOCATION with the message Location not found.

UpdateActiveBreakpointRequest
Request to update an active breakpoint.

UpdateActiveBreakpointResponse
Response for updating an active breakpoint. The message is defined to allow future extensions.

Variable
Represents a variable or an argument possibly of a compound object type. Note how the following variables are represented:

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.