Struct google_clouddebugger2::Breakpoint

pub struct Breakpoint {
    pub status: Option<StatusMessage>,
    pub variable_table: Option<Vec<Variable>>,
    pub user_email: Option<String>,
    pub log_level: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub final_time: Option<String>,
    pub stack_frames: Option<Vec<StackFrame>>,
    pub expressions: Option<Vec<String>>,
    pub evaluated_expressions: Option<Vec<Variable>>,
    pub id: Option<String>,
    pub condition: Option<String>,
    pub log_message_format: Option<String>,
    pub create_time: Option<String>,
    pub location: Option<SourceLocation>,
    pub action: Option<String>,
    pub is_final_state: Option<bool>,
}

Represents the breakpoint specification, status and results.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

debuggees breakpoints set debugger (request)

Fields

status

Breakpoint status. The status includes an error flag and a human readable message. This field is usually unset. The message can be either informational or an error message. Regardless, clients should always display the text message back to the user. Error status indicates complete failure of the breakpoint. Example (non-final state): Still loading symbols... Examples (final state): * Invalid line number referring to location * Field f not found in class C referring to condition

variable_table

The variable_table exists to aid with computation, memory and network traffic optimization. It enables storing a variable once and reference it from multiple variables, including variables stored in the variable_table itself. For example, the same this object, which may appear at many levels of the stack, can have all of its data stored once in this table. The stack frame variables then would hold only a reference to it. The variable var_table_index field is an index into this repeated field. The stored objects are nameless and get their name from the referencing variable. The effective variable is a merge of the referencing variable and the referenced variable.

user_email

E-mail address of the user that created this breakpoint

log_level

Indicates the severity of the log. Only relevant when action is LOG.

labels

A set of custom breakpoint properties, populated by the agent, to be displayed to the user.

final_time

Time this breakpoint was finalized as seen by the server in seconds resolution.

stack_frames

The stack at breakpoint time.

expressions

List of read-only expressions to evaluate at the breakpoint location. The expressions are composed using expressions in the programming language at the source location. If the breakpoint action is LOG, the evaluated expressions are included in log statements.

evaluated_expressions

Values of evaluated expressions at breakpoint time. The evaluated expressions appear in exactly the same order they are listed in the expressions field. The name field holds the original expression text, the value or members field holds the result of the evaluated expression. If the expression cannot be evaluated, the status inside the Variable will indicate an error and contain the error text.

id

Breakpoint identifier, unique in the scope of the debuggee.

condition

Condition that triggers the breakpoint. The condition is a compound boolean expression composed using expressions in a programming language at the source location.

log_message_format

Only relevant when action is LOG. Defines the message to log when the breakpoint hits. The message may include parameter placeholders $0, $1, etc. These placeholders are replaced with the evaluated value of the appropriate expression. Expressions not referenced in log_message_format are not logged. Example: Message received, id = $0, count = $1 with expressions = [ message.id, message.count ].

create_time

Time this breakpoint was created by the server in seconds resolution.

location

Breakpoint source location.

action

Action that the agent should perform when the code at the breakpoint location is hit.

is_final_state

When true, indicates that this is a final result and the breakpoint state will not change from here on.