Struct google_clouddebugger2::CloudRepoSourceContext

pub struct CloudRepoSourceContext {
    pub alias_context: Option<AliasContext>,
    pub revision_id: Option<String>,
    pub alias_name: Option<String>,
    pub repo_id: Option<RepoId>,
}

A CloudRepoSourceContext denotes a particular revision in a cloud repo (a repo hosted by the Google Cloud Platform).

This type is not used in any activity, and only used as part of another schema.

Fields

alias_context

An alias, which may be a branch or tag.

revision_id

A revision ID.

alias_name

The name of an alias (branch, tag, etc.).

repo_id

The ID of the repo.