Struct google_clouddebugger2::CloudWorkspaceId

pub struct CloudWorkspaceId {
    pub repo_id: Option<RepoId>,
    pub name: Option<String>,
}

A CloudWorkspaceId is a unique identifier for a cloud workspace. A cloud workspace is a place associated with a repo where modified files can be stored before they are committed.

This type is not used in any activity, and only used as part of another schema.

Fields

repo_id

The ID of the repo containing the workspace.

name

The unique name of the workspace within the repo. This is the name chosen by the client in the Source API's CreateWorkspace method.