Struct google_clouddebugger2::Debuggee

pub struct Debuggee {
    pub status: Option<StatusMessage>,
    pub description: Option<String>,
    pub is_disabled: Option<bool>,
    pub labels: Option<HashMap<String, String>>,
    pub uniquifier: Option<String>,
    pub project: Option<String>,
    pub source_contexts: Option<Vec<SourceContext>>,
    pub ext_source_contexts: Option<Vec<ExtendedSourceContext>>,
    pub agent_version: Option<String>,
    pub is_inactive: Option<bool>,
    pub id: Option<String>,
}

Represents the application to debug. The application may include one or more replicated processes executing the same code. Each of these processes is attached with a debugger agent, carrying out the debugging commands. The agents attached to the same debuggee are identified by using exactly the same field values when registering.

This type is not used in any activity, and only used as part of another schema.

Fields
status: Option<StatusMessage>

Human readable message to be displayed to the user about this debuggee. Absence of this field indicates no status. The message can be either informational or an error status.

description: Option<String>

Human readable description of the debuggee. Including a human-readable project name, environment name and version information is recommended.

is_disabled: Option<bool>

If set to true, indicates that the agent should disable itself and detach from the debuggee.

labels: Option<HashMap<String, String>>

A set of custom debuggee properties, populated by the agent, to be displayed to the user.

uniquifier: Option<String>

Debuggee uniquifier within the project. Any string that identifies the application within the project can be used. Including environment and version or build IDs is recommended.

project: Option<String>

Project the debuggee is associated with. Use the project number when registering a Google Cloud Platform project.

source_contexts: Option<Vec<SourceContext>>

References to the locations and revisions of the source code used in the deployed application. NOTE: This field is deprecated. Consumers should use ext_source_contexts if it is not empty. Debug agents should populate both this field and ext_source_contexts.

ext_source_contexts: Option<Vec<ExtendedSourceContext>>

References to the locations and revisions of the source code used in the deployed application. Contexts describing a remote repo related to the source code have a category label of remote_repo. Source snapshot source contexts have a category of snapshot.

agent_version: Option<String>

Version ID of the agent release. The version ID is structured as following: domain/type/vmajor.minor (for example google.com/gcp-java/v1.1).

is_inactive: Option<bool>

If set to true, indicates that the debuggee is considered as inactive by the Controller service.

id: Option<String>

Unique identifier for the debuggee generated by the controller service.