Struct google_clouddebugger2::GerritSourceContext

pub struct GerritSourceContext {
    pub alias_context: Option<AliasContext>,
    pub revision_id: Option<String>,
    pub host_uri: Option<String>,
    pub alias_name: Option<String>,
    pub gerrit_project: Option<String>,
}

A SourceContext referring to a Gerrit project.

This type is not used in any activity, and only used as part of another schema.

Fields

alias_context

An alias, which may be a branch or tag.

revision_id

A revision (commit) ID.

host_uri

The URI of a running Gerrit instance.

alias_name

The name of an alias (branch, tag, etc.).

gerrit_project

The full project name within the host. Projects may be nested, so "project/subproject" is a valid project name. The "repo name" is hostURI/project.