Struct google_clouddebugger2::ListActiveBreakpointsResponse

pub struct ListActiveBreakpointsResponse {
    pub wait_expired: Option<bool>,
    pub next_wait_token: Option<String>,
    pub breakpoints: Option<Vec<Breakpoint>>,
}

Response for listing active breakpoints.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

debuggees breakpoints list controller (response)

Fields

wait_expired

The wait_expired field is set to true by the server when the request times out and the field success_on_timeout is set to true.

next_wait_token

A wait token that can be used in the next method call to block until the list of breakpoints changes.

breakpoints

List of all active breakpoints. The fields id and location are guaranteed to be set on each breakpoint.