Struct google_clouddebugger2::StatusMessage

pub struct StatusMessage {
    pub is_error: Option<bool>,
    pub refers_to: Option<String>,
    pub description: Option<FormatMessage>,
}

Represents a contextual status message. The message can indicate an error or informational status, and refer to specific parts of the containing object. For example, the Breakpoint.status field can indicate an error referring to the BREAKPOINT_SOURCE_LOCATION with the message Location not found.

This type is not used in any activity, and only used as part of another schema.

Fields

is_error

Distinguishes errors from informational messages.

refers_to

Reference to which the message applies.

description

Status message text.